Added description and how to setup and use project
Provided details covering the overview of the project,
contents of the repository, project dependencies and
steps on how to setup the project anaconda environment
that uses tensorflow 2.0 and openAI gym, specifically
the super mario bros gym interface. Instructions for
running the demo are in the works, but first I need
to build the reinforcement learning appliction.